Implement ABD
This commit introduces a first draft of the Generic ABD implementation

The local storage is modeled by the abd::store::LocalStore trait, which
may (or not) implement persistence.

Client ids are first-order entities.

The timestamp is generic, taking as a type parameter the (generic)
client id and another type, as long as it is Ord, Eq, and Min (meaning,
the smallest value of the type is known).

When connection errors cause quorums to become unavailable, those errors
are propagated.